Change player_score type from f64 to f32
Modified the player_score field in the MMRHistory struct to use f32 instead of f64, reducing memory usage while maintaining sufficient precision for the application's needs. This change optimizes performance and aligns with similar fields in the codebase.

+++ src/routes/v1/players/mmr_history.rs
@@ -26,7 +26,7 @@ pub struct HeroMMRHistoryQuery {
 pub struct MMRHistory {
     pub match_id: u64,
     /// Player Score is the index for the rank array (internally used for the rank regression)
-    pub player_score: f64,
+    pub player_score: f32,
     /// The Player Rank
     pub rank: u32,
     /// Extracted from the rank the division (rank // 10)
